pyproject.toml
[project]
name = "image-host"
version = "0.1.0"
description = "Small image hosting service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==6.0.3",
    "django-browser-reload==1.21.0",
    "django-vite==3.1.0",
    "djangorestframework==3.17.1",
]

[dependency-groups]
dev = [
    "mypy==1.20.0",
    "djangorestframework-stubs[compatible-mypy]==3.16.9",
    "django-stubs[compatible-mypy]==6.0.2",
]

[tool.mypy]
strict = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "project.settings"
